решены 21,25,28,29,30   в работе 26,31

+++ src/lab6/Task16.java
@@ -5,7 +5,7 @@ public class Task16 {
     public static void main(String[] args) {
 
         double a = 20;
-        int b = 40;
+        double b = 40;
         double c = 21;
         int a3 = 3;
